pyproject.toml
[project]
name = "natuurgegevens-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "daanknoors",email = "d.knoors@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.0,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "jupyterlab (>=4.4.3,<5.0.0)",
    "geopandas (>=1.1.1,<2.0.0)",
    "contextily (>=1.6.2,<2.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyjanitor (>=0.32.1,<0.33.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
